requirements.txt
streamlit
pytesseract
pillow
pandas
openpyxl
PyMuPDF
cryptography
fastapi
uvicorn
